HHH-18990 move Path and Restriction to their own package
