+++ hibernate-core/src/main/java/org/hibernate/query/restriction/NamedAttributeRange.java
@@ -2,7 +2,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later
  * Copyright Red Hat Inc. and Hibernate Authors
  */
-package org.hibernate.query;
+package org.hibernate.query.restriction;
 
 import jakarta.persistence.criteria.CriteriaBuilder;
 import jakarta.persistence.criteria.Predicate;
