+++ hibernate-core/src/main/java/org/hibernate/query/restriction/Negation.java
@@ -2,29 +2,28 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later
  * Copyright Red Hat Inc. and Hibernate Authors
  */
-package org.hibernate.query;
+package org.hibernate.query.restriction;
 
 import jakarta.persistence.criteria.CriteriaBuilder;
 import jakarta.persistence.criteria.Predicate;
 import jakarta.persistence.criteria.Root;
-import org.hibernate.query.range.Range;
 
 /**
- * Restricts a path from an entity to a given {@link Range}.
+ * Negates a restriction; a logical NOT.
  *
+ * @param restriction The restriction to be negated
  * @param <X> The entity type
- * @param <U> The attribute type
  *
  * @author Gavin King
  */
-record PathRange<X, U>(Path<X, U> path, Range<? super U> range) implements Restriction<X> {
+record Negation<X>(Restriction<X> restriction) implements Restriction<X> {
 	@Override
 	public Restriction<X> negated() {
-		return new Negation<>( this );
+		return restriction;
 	}
 
 	@Override
 	public Predicate toPredicate(Root<? extends X> root, CriteriaBuilder builder) {
-		return range.toPredicate( path.path( root ), builder );
+		return builder.not( restriction.toPredicate( root, builder ) );
 	}
 }
