+++ hibernate-core/src/main/java/org/hibernate/query/restriction/PathRange.java
@@ -2,23 +2,22 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later
  * Copyright Red Hat Inc. and Hibernate Authors
  */
-package org.hibernate.query;
+package org.hibernate.query.restriction;
 
 import jakarta.persistence.criteria.CriteriaBuilder;
 import jakarta.persistence.criteria.Predicate;
 import jakarta.persistence.criteria.Root;
-import jakarta.persistence.metamodel.SingularAttribute;
 import org.hibernate.query.range.Range;
 
 /**
- * Restricts an attribute of an entity to a given {@link Range}.
+ * Restricts a path from an entity to a given {@link Range}.
  *
  * @param <X> The entity type
  * @param <U> The attribute type
  *
  * @author Gavin King
  */
-record AttributeRange<X, U>(SingularAttribute<X, U> attribute, Range<U> range) implements Restriction<X> {
+record PathRange<X, U>(Path<X, U> path, Range<? super U> range) implements Restriction<X> {
 	@Override
 	public Restriction<X> negated() {
 		return new Negation<>( this );
@@ -26,6 +25,6 @@ record AttributeRange<X, U>(SingularAttribute<X, U> attribute, Range<U> range) i
 
 	@Override
 	public Predicate toPredicate(Root<? extends X> root, CriteriaBuilder builder) {
-		return range.toPredicate( root.get( attribute ), builder );
+		return range.toPredicate( path.path( root ), builder );
 	}
 }
