+++ hibernate-core/src/test/java/org/hibernate/orm/test/query/restriction/RestrictionTest.java
@@ -18,19 +18,19 @@ import java.util.List;
 
 import static org.hibernate.query.Order.asc;
 import static org.hibernate.query.Order.desc;
-import static org.hibernate.query.Path.from;
-import static org.hibernate.query.Restriction.all;
-import static org.hibernate.query.Restriction.any;
-import static org.hibernate.query.Restriction.between;
-import static org.hibernate.query.Restriction.contains;
-import static org.hibernate.query.Restriction.endsWith;
-import static org.hibernate.query.Restriction.equal;
-import static org.hibernate.query.Restriction.equalIgnoringCase;
-import static org.hibernate.query.Restriction.greaterThan;
-import static org.hibernate.query.Restriction.in;
-import static org.hibernate.query.Restriction.like;
-import static org.hibernate.query.Restriction.restrict;
-import static org.hibernate.query.Restriction.unrestricted;
+import static org.hibernate.query.restriction.Path.from;
+import static org.hibernate.query.restriction.Restriction.all;
+import static org.hibernate.query.restriction.Restriction.any;
+import static org.hibernate.query.restriction.Restriction.between;
+import static org.hibernate.query.restriction.Restriction.contains;
+import static org.hibernate.query.restriction.Restriction.endsWith;
+import static org.hibernate.query.restriction.Restriction.equal;
+import static org.hibernate.query.restriction.Restriction.equalIgnoringCase;
+import static org.hibernate.query.restriction.Restriction.greaterThan;
+import static org.hibernate.query.restriction.Restriction.in;
+import static org.hibernate.query.restriction.Restriction.like;
+import static org.hibernate.query.restriction.Restriction.restrict;
+import static org.hibernate.query.restriction.Restriction.unrestricted;
 import static org.hibernate.query.range.Range.containing;
 import static org.hibernate.query.range.Range.greaterThan;
 import static org.hibernate.query.range.Range.singleCaseInsensitiveValue;
