+++ tooling/metamodel-generator/src/main/java/org/hibernate/processor/util/Constants.java
@@ -91,7 +91,7 @@ public final class Constants {
 	public static final String HIB_KEYED_PAGE = "org.hibernate.query.KeyedPage";
 	public static final String HIB_KEYED_RESULT_LIST = "org.hibernate.query.KeyedResultList";
 	public static final String HIB_SORT_DIRECTION = "org.hibernate.query.SortDirection";
-	public static final String HIB_RESTRICTION = "org.hibernate.query.Restriction";
+	public static final String HIB_RESTRICTION = "org.hibernate.query.restriction.Restriction";
 	public static final String HIB_RANGE = "org.hibernate.query.range.Range";
 
 	public static final String CHECK_HQL = "org.hibernate.annotations.processing.CheckHQL";
